# examhall/openapi.yaml
...
openapi: 3.0.3
info:
  title: ELEVIX - 1.0 Fullstack App Contract
  version: 1.0.0
  description: >
    Locked API contract for PS F1, F3, F4, F5, F7 and F8. Implement every
    path. The core logic must be handwritten - no auto-CRUD framework may
    generate these routes for you.
servers:
  - url: http://localhost:{port}
    variables:
      port:
        default: "8080"

# ---------------------------------------------------------------------------
# Every app must answer this, whatever the PS.
# ---------------------------------------------------------------------------
paths:
...
  /items:
    get:
      summary: F1 - search found items.
      operationId: f1_search_items
      security: []
      parameters:
        - {name: q,      in: query, required: false, schema: {type: string}}
        - {name: status, in: query, required: false,
           schema: {type: string, enum: [open, claimed, closed]}}
        - {name: page,   in: query, required: false, schema: {type: integer, default: 1}}
        - {name: limit,  in: query, required: false, schema: {type: integer, default: 20}}
      responses:
        "200":
          description: Matching items.
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items: {$ref: "#/components/schemas/F1Item"}
                  total: {type: integer}
    post:
      summary: F1 - report a found item.
      operationId: f1_create_item
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/F1Item"}
        "400": {$ref: "#/components/responses/BadRequest"}
  /items/{id}/claim:
    post:
      summary: F1 - claim an item; the handwritten matching logic runs here.
      operationId: f1_claim_item
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [claimant, proof]
              properties:
                claimant: {type: string}
                proof:    {type: string, description: Answer to the security question.}
      responses:
        "200":
          description: Claim accepted.
          content:
            application/json:
              schema:
                type: object
                required: [id, status]
                properties:
                  id: {type: string}
                  status: {type: string, enum: [claimed]}
        "403":
          description: Proof did not match.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "404": {$ref: "#/components/responses/NotFound"}

# ---------------------------------------------------------------------------
# F3  Event Registration + E-Certificate Portal
# ---------------------------------------------------------------------------
  /events:
    get:
      summary: F3 - list events.
      operationId: f3_list_events
      security: []
      responses:
        "200":
          description: Events.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: "#/components/schemas/F3Event"}
    post:
      summary: F3 - create an event.
      operationId: f3_create_event
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/F3Event"}
        "400": {$ref: "#/components/responses/BadRequest"}
  /events/{id}/register:
    post:
      summary: F3 - register a participant.
      operationId: f3_register
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email]
              properties:
                name:  {type: string}
                email: {type: string, format: email}
                dept:  {type: string}
      responses:
        "201":
          description: Registered.
          content:
            application/json:
              schema:
                type: object
                required: [reg_id, certificate_id]
                properties:
                  reg_id:        {type: string}
                  certificate_id: {type: string, description: The handwritten certificate ID.}
        "400": {$ref: "#/components/responses/BadRequest"}
        "409":
          description: Already registered.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
  /certificates/{certificate_id}:
    get:
      summary: F3 - verify a certificate by its ID.
      operationId: f3_verify_certificate
      security: []
      parameters:
        - {name: certificate_id, in: path, required: true, schema: {type: string}}
      responses:
        "200":
          description: Certificate is genuine.
          content:
            application/json:
              schema:
                type: object
                required: [certificate_id, valid, name, event]
                properties:
                  certificate_id: {type: string}
                  valid:          {type: boolean}
                  name:           {type: string}
                  event:          {type: string}
                  issued_on:      {type: string, format: date}
        "404": {$ref: "#/components/responses/NotFound"}

# ---------------------------------------------------------------------------
# F4  Lab Equipment Booking System
# ---------------------------------------------------------------------------
  /equipment:
    get:
      summary: F4 - list bookable equipment.
      operationId: f4_list_equipment
      security: []
      responses:
        "200":
          description: Equipment list.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: "#/components/schemas/F4Equipment"}
    post:
      summary: F4 - add equipment.
      operationId: f4_create_equipment
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/F4Equipment"}
        "400": {$ref: "#/components/responses/BadRequest"}
  /bookings:
    post:
      summary: F4 - request a booking; slot-conflict detection runs here.
      operationId: f4_create_booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [equipment_id, user, from, to]
              properties:
                equipment_id: {type: string}
                user:         {type: string}
                from:         {type: string, format: date-time}
                to:           {type: string, format: date-time}
      responses:
        "201":
          description: Booking accepted.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/F4Booking"}
        "409":
          description: Slot clash - must be rejected, never silently overwritten.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "400": {$ref: "#/components/responses/BadRequest"}
  /bookings/{id}/approve:
    post:
      summary: F4 - move a booking through the approval state machine.
      operationId: f4_approve_booking
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        "200":
          description: New state.
          content:
            application/json:
              schema:
                type: object
                required: [id, status]
                properties:
                  id:     {type: string}
                  status: {type: string, enum: [pending, approved, rejected, returned]}
        "404": {$ref: "#/components/responses/NotFound"}

# ---------------------------------------------------------------------------
# F5  Hostel Complaint Tracker
# ---------------------------------------------------------------------------
  /complaints:
    get:
      summary: F5 - list complaints.
      operationId: f5_list_complaints
      security: []
      parameters:
        - {name: status, in: query, required: false,
           schema: {type: string, enum: [open, in_progress, escalated, closed]}}
        - {name: hostel, in: query, required: false, schema: {type: string}}
      responses:
        "200":
          description: Complaints.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: "#/components/schemas/F5Complaint"}
    post:
      summary: F5 - file a complaint.
      operationId: f5_create_complaint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [student, hostel, room, category, text]
              properties:
                student:  {type: string}
                hostel:   {type: string}
                room:     {type: string}
                category: {type: string, enum: [electrical, plumbing, cleaning, other]}
                text:     {type: string}
                priority: {type: string, enum: [low, normal, high], default: normal}
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/F5Complaint"}
        "400": {$ref: "#/components/responses/BadRequest"}
  /complaints/{id}/status:
    patch:
      summary: F5 - change status; the SLA timer and escalation rules run here.
      operationId: f5_update_status
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: {type: string, enum: [open, in_progress, escalated, closed]}
                note:   {type: string}
      responses:
        "200":
          description: Updated, with the SLA fields recalculated.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/F5Complaint"}
        "404": {$ref: "#/components/responses/NotFound"}

# ---------------------------------------------------------------------------
# F7  Exam Hall Allocation Tool
# ---------------------------------------------------------------------------
...
  /students:
    get:
      summary: F7 - list students to be seated.
      operationId: f7_list_students
      security: []
      parameters:
        - {name: dept, in: query, required: false, schema: {type: string}}
        - {name: year, in: query, required: false, schema: {type: integer}}
      responses:
        "200":
          description: Students.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: "#/components/schemas/F7Student"}
  /api/config:
    get:
      summary: F7 extension - retrieve the saved seating configuration.
      operationId: f7_get_config
      security: []
      responses:
        "200":
          description: Current groups, halls, and allocation settings.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/F7Config"}
    post:
      summary: F7 extension - save student groups and hall details.
      operationId: f7_save_config
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/F7Config"}
      responses:
        "200":
          description: Configuration saved.
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok: {type: boolean}
        "400": {$ref: "#/components/responses/BadRequest"}
  /allocate:
    post:
      summary: F7 - run the constraint-satisfaction allocator. No OR-tools allowed.
      operationId: f7_allocate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [exam_id]
              properties:
                exam_id: {type: string}
                rules:
                  type: object
                  properties:
                    same_dept_adjacent: {type: boolean, default: false}
                    max_per_row:        {type: integer, default: 2}
      responses:
        "200":
          description: Allocation result.
          content:
            application/json:
              schema:
                type: object
                required: [exam_id, allocations, unseated]
                properties:
                  exam_id:     {type: string}
                  allocations:
                    type: array
                    items:
                      type: object
                      required: [hall, seat, student]
                      properties:
                        hall:    {type: string}
                        seat:    {type: string}
                        student: {type: string}
                  unseated:
                    type: array
                    items: {type: string}
        "400": {$ref: "#/components/responses/BadRequest"}
  /api/allocate:
    post:
      summary: F7 extension - allocate the saved configuration for the web application.
      operationId: f7_allocate_saved_config
      responses:
        "200":
          description: Seating plan generated from the saved configuration.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/F7SeatingPlan"}
        "422":
          description: No valid arrangement can be generated.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
...
components:
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
          example: {error: BAD_REQUEST}
    NotFound:
      description: No such resource.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
          example: {error: NOT_FOUND}
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:   {type: string, description: A short machine-readable code.}
        message: {type: string, description: A human-readable explanation.}
    F1Item:
      type: object
      required: [id, title, status]
      properties:
        id:         {type: string}
        title:      {type: string}
        found_at:   {type: string}
        found_on:   {type: string, format: date}
        status:     {type: string, enum: [open, claimed, closed]}
        question:   {type: string, description: Security question for the claimant.}
    F3Event:
      type: object
      required: [id, title, date]
      properties:
        id:    {type: string}
        title: {type: string}
        date:  {type: string, format: date}
        venue: {type: string}
    F4Equipment:
      type: object
      required: [id, name, quantity]
      properties:
        id:       {type: string}
        name:     {type: string}
        quantity: {type: integer}
...
security: []